MOON-321: Switch controlled/uncontrolled naming and other improvements
