+++ src/components/Input/Input.spec.js
@@ -55,6 +55,14 @@ describe('Input', () => {
         expect(screen.getByRole('search')).toBeInTheDocument();
     });
 
+    it('should work when no value or defaultValue is specified', () => {
+        render(<Input data-testid="moonstone-input"/>);
+        userEvent.type(screen.getByTestId('moonstone-input'), 'type a value');
+        expect(screen.getByDisplayValue('type a value')).toBeInTheDocument();
+    });
+});
+
+describe('UncontrolledInput', () => {
     it('should have specified defaultValue', () => {
         render(<Input defaultValue="test-default-value"/>);
         expect(screen.getByDisplayValue('test-default-value')).toBeInTheDocument();
@@ -66,17 +74,38 @@ describe('Input', () => {
         expect(screen.getByDisplayValue('test-default-value-updated')).toBeInTheDocument();
     });
 
+    it('should reset field when we click on the reset button of the search input', () => {
+        render(<Input data-testid="moonstone-input" defaultValue="test-default-value" variant="search"/>);
+        userEvent.click(screen.getByLabelText('Reset'));
+        expect(screen.getByTestId('moonstone-input')).toHaveValue('');
+    });
+
+    it('should call specified onChange function', () => {
+        const handleChange = jest.fn();
+
+        render(<Input data-testid="moonstone-input" defaultValue="test-default-value" onChange={handleChange}/>);
+        userEvent.type(screen.getByTestId('moonstone-input'), '1');
+
+        expect(handleChange).toHaveBeenCalledTimes(1);
+    });
+
+    it('should call specified onClear function', () => {
+        const handleClear = jest.fn();
+
+        render(<Input variant="search" defaultValue="test-default-value" onClear={handleClear}/>);
+        userEvent.click(screen.getByLabelText('Reset'));
+
+        expect(handleClear).toHaveBeenCalledTimes(1);
+    });
+});
+
+describe('ControlledInput', () => {
     it('should display specified value', () => {
         render(<Input value="test-value" onChange={() => {}}/>);
         expect(screen.getByDisplayValue('test-value')).toBeInTheDocument();
     });
 
-    it('should display specified value', () => {
-        render(<Input value="test-value" onChange={() => {}}/>);
-        expect(screen.getByDisplayValue('test-value')).toBeInTheDocument();
-    });
-
-    it('should call the onChange function when we use the controlled version', () => {
+    it('should call specified onChange function', () => {
         const handleChange = jest.fn();
 
         render(<Input data-testid="moonstone-input" value="test-value" onChange={handleChange}/>);
@@ -85,12 +114,12 @@ describe('Input', () => {
         expect(handleChange).toHaveBeenCalledTimes(1);
     });
 
-    it('should call the onChange function when we use the uncontrolled version', () => {
-        const handleChange = jest.fn();
+    it('should call specified onClear function', () => {
+        const handleClear = jest.fn();
 
-        render(<Input data-testid="moonstone-input" defaultValue="test-value" onChange={handleChange}/>);
-        userEvent.type(screen.getByTestId('moonstone-input'), '1');
+        render(<Input variant="search" value="test-value" onChange={() => {}} onClear={handleClear}/>);
+        userEvent.click(screen.getByLabelText('Reset'));
 
-        expect(handleChange).toHaveBeenCalledTimes(1);
+        expect(handleClear).toHaveBeenCalledTimes(1);
     });
 });
